backend/requirements.txt
Django==1.6.5
djangorestframework==2.3.13
psycopg2==2.5.3
wsgiref==0.1.2
